fix: Fix dependency on DH API classes and rename some DH compat classes
