+++ src/main/java/io/github/steveplays28/blendium/client/compat/distanthorizons/BlendiumAfterDhInitEventHandler.java
@@ -1,4 +1,4 @@
-package io.github.steveplays28.blendium.client.compat;
+package io.github.steveplays28.blendium.client.compat.distanthorizons;
 
 import com.seibel.distanthorizons.api.DhApi;
 import com.seibel.distanthorizons.api.methods.events.abstractEvents.DhApiAfterDhInitEvent;
@@ -7,14 +7,14 @@ import net.coderbot.iris.Iris;
 import net.fabricmc.loader.api.FabricLoader;
 
 import static io.github.steveplays28.blendium.client.BlendiumClient.*;
-import static io.github.steveplays28.blendium.client.compat.iris.BlendiumDHShaderpackPresets.applyDHShaderpackPreset;
+import static io.github.steveplays28.blendium.client.compat.iris.BlendiumDhShaderpackPresets.applyDhShaderpackPreset;
 
 // TODO: Maybe reset the LOD shaderpack presets when unloading the game, in case Iris gets disabled/removed
 public class BlendiumAfterDhInitEventHandler extends DhApiAfterDhInitEvent {
 	@Override
 	public void afterDistantHorizonsInit(DhApiEventParam<Void> input) {
 		if (FabricLoader.getInstance().isModLoaded(IRIS_SHADERS_MOD_ID)) {
-			applyDHShaderpackPreset(Iris.getCurrentPackName());
+			applyDhShaderpackPreset(Iris.getCurrentPackName());
 		}
 
 		DhApi.Delayed.configs.graphics().brightnessMultiplier().addChangeListener(this::onBrightnessMultiplierChanged);
